Vectorfieldvisualizationgl: update to new module structure
