+++ modules/vectorfieldvisualizationgl/include/modules/vectorfieldvisualizationgl/processors/2d/vector2dcurl.h
