+++ modules/vectorfieldvisualizationgl/include/modules/vectorfieldvisualizationgl/processors/3d/vector3ddivergence.h
