+++ modules/vectorfieldvisualizationgl/src/processors/2d/vector2dmagnitude.cpp
@@ -27,7 +27,7 @@
  *
  *********************************************************************************/
 
-#include "vector2dcurl.h"
+#include <modules/vectorfieldvisualizationgl/processors/2d/vector2dmagnitude.h>
 #include <modules/opengl/texture/textureunit.h>
 #include <modules/opengl/texture/textureutils.h>
 #include <modules/opengl/image/imagegl.h>
@@ -36,29 +36,28 @@
 namespace inviwo {
 
 // The Class Identifier has to be globally unique. Use a reverse DNS naming scheme
-const ProcessorInfo Vector2DCurl::processorInfo_{
-    "org.inviwo.Vector2DCurl",     // Class identifier
-    "Vector 2D Curl",              // Display name
-    "Vector Field Visualization",  // Category
-    CodeState::Experimental,       // Code state
-    Tags::GL,                      // Tags
+const ProcessorInfo Vector2DMagnitude::processorInfo_{
+    "org.inviwo.Vector2DMagnitude",  // Class identifier
+    "Vector 2D Magnitude",           // Display name
+    "Vector Field Visualization",    // Category
+    CodeState::Stable,               // Code state
+    Tags::GL,                        // Tags
 };
-const ProcessorInfo Vector2DCurl::getProcessorInfo() const { return processorInfo_; }
+const ProcessorInfo Vector2DMagnitude::getProcessorInfo() const { return processorInfo_; }
 
-Vector2DCurl::Vector2DCurl()
+Vector2DMagnitude::Vector2DMagnitude()
     : Processor()
     , inport_("inport", true)
-    , outport_("outport", DataVec4Float32::get())
-    , shader_("vector2dcurl.frag")
-
-{
+    , outport_("outport", DataFloat32::get())
+    , shader_("vector2dmagnitude.frag") {
 
     addPort(inport_);
     addPort(outport_);
 }
 
-void Vector2DCurl::process() {
+void Vector2DMagnitude::process() {
     utilgl::activateAndClearTarget(outport_);
+    outport_.getEditableData()->getColorLayer()->setSwizzleMask(swizzlemasks::luminance);
 
     shader_.activate();
     TextureUnitContainer units;
