+++ modules/vectorfieldvisualizationgl/src/processors/datageneration/lorenzsystem.cpp
@@ -27,7 +27,7 @@
  *
  *********************************************************************************/
 
-#include "lorenzsystem.h"
+#include <modules/vectorfieldvisualizationgl/processors/datageneration/lorenzsystem.h>
 #include <modules/opengl/texture/textureunit.h>
 #include <modules/opengl/texture/textureutils.h>
 #include <modules/opengl/shader/shaderutils.h>
