+++ modules/vectorfieldvisualizationgl/src/vectorfieldvisualizationglmodule.cpp
